requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
camelot-py==1.0.9
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.0.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==3.17.4
pypdfium2==5.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
six==1.17.0
starlette==0.49.3
tabulate==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.39.0
